refactor(console): 拆分内容治理列与手动动作区
- 抽出举报队列和治理日志列定义工厂

- 抽出手动治理动作区与生效状态卡片

- 保持内容治理 API、表单和动作语义不变
